+++ riscv/insns/bnec.h
@@ -1,2 +1,2 @@
-if (RS1 != insn.ade_bec_cimm())
+if ((int64_t)RS1 != insn.ade_bec_cimm())
     set_pc(BBCBRANCH_TARGET);
